pyproject.toml
[project]
name = "repowiki"
version = "0.1.0"
description = "Hierarchical wiki generator for code repositories using LightRAG knowledge graphs"
authors = [
    {name = "LightRAG Contributors"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "networkx>=3.0.0",
    "nano-vectordb>=0.0.4",
    "python-dotenv>=1.0.0",
    "pipmaster>=1.1.0",
    # LlamaIndex and LiteLLM for GitHub Copilot support
    "llama-index-core>=0.10.0",
    "llama-index-llms-litellm>=0.1.0",
    "llama-index-embeddings-litellm>=0.1.0",
    # Use forked litellm with GitHub Copilot support
    "litellm @ git+https://github.com/fenghaitao/litellm.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
repowiki = "repowiki.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py38"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
